Updated PTKSwitchToGitHub

+++ PTKSwitchToGitHub.m
@@ -1,4 +1,4 @@
-function PTKSwitchToGitHub
+function PTKSwitchToGitHub(varargin)
 
     % PTKSwitchToGitHub. A script to switch a Google Code svn checkout to a
     %     GitHub git checkout
@@ -12,25 +12,29 @@ function PTKSwitchToGitHub
     %     Distributed under the GNU GPL v3 licence. Please see website for details.
     %
 
-    
-    if isGitInstalled
-        answer = questdlg('PTK has moved to GitHub. Should I migrate your PTK codebase now?','PTK has moved to GitHub','Do not migrate','Migrate','Migrate');
-        if strcmp(answer, 'Migrate')
-            answer2 = questdlg('Please ensure you have any important changes backed up. If successful, the migration should preserve any local changes you have made to PTK, but I cannot guarantee this!','Pulmonary Toolkit','Cancel migration','Migrate','Migrate');
-            if strcmp(answer2, 'Migrate')
-                if SwitchToGitHub
-                    msgbox('Successfully migrated to GitHub. Please now use Git to pull updates to PTK', 'Migrated to GitHub');
-                else
-                    msgbox('Sorry, there was a problem migrating your codebase. Please clone the new codebae yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Failed to migrate to GitHub');
+    if ~checkIfGitRepoExists
+        if ~checkDoNotUpdate || (nargin > 0 && strcmp(varargin{1}, 'force'))
+            clearDoNotUpdate;
+            if isGitInstalled
+                answer = questdlg('PTK has moved to GitHub. Should I migrate your PTK codebase now?','PTK has moved to GitHub','Later','Do not ask me again', 'Migrate','Migrate');
+                if strcmp(answer, 'Do not ask me again')
+                    setDoNotUpdateFlag
+                elseif strcmp(answer, 'Migrate')
+                    answer2 = questdlg('Please ensure you have any important changes backed up. If successful, the migration should preserve any local changes you have made to PTK, but I cannot guarantee this!','Pulmonary Toolkit','Cancel migration','Migrate','Migrate');
+                    if strcmp(answer2, 'Migrate')
+                        if SwitchToGitHub
+                            msgbox('Successfully migrated to GitHub. Please now use Git to pull updates to PTK', 'Migrated to GitHub');
+                        else
+                            msgbox('Sorry, there was a problem migrating your codebase. Please clone the new codebae yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Failed to migrate to GitHub');
+                        end
+                    else
+                        msgbox('Please re-run ptk to migrate your codebase, or clone the new codebae yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Pulmonary Toolkit');
+                    end
                 end
             else
-                msgbox('Please re-run ptk to migrate your codebase, or clone the new codebae yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Pulmonary Toolkit');
+                msgbox('PTK has moved to GitHub. I could not find Git on your path. Please install Git and re-run and I will try to update your codebase. Alternatively, clone PTK yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Pulmonary Toolkit');
             end
-        else
-            msgbox('Please re-run ptk to migrate your codebase, or clone the new codebae yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Pulmonary Toolkit');
         end
-    else
-        msgbox('PTK has moved to GitHub. I could not find Git on your path. Please install Git and re-run and I will try to update your codebase. Alternatively, clone PTK yourself from https://github.com/tomdoel/pulmonarytoolkit', 'Pulmonary Toolkit');
     end
 end
 
@@ -86,6 +90,11 @@ function success = SwitchToGitHub
         return;
     end
     
+    if ~execute('git checkout -b master')
+        disp('Could not checkout the master branch');
+        return;
+    end
+    
     success = true;
 end
 
@@ -94,7 +103,7 @@ function installed = isGitInstalled
 end
 
 function success = execute(command)
-    return_value = system(command);
+    [return_value, ~] = system(command);
     success = return_value == 0;
 end
 
@@ -107,4 +116,35 @@ function fixCurlPath
     if (7 == exist(binDir, 'dir')) && ~strcmp(currentLibPath(1:length(binDir) + 1), [binDir ':'])
         setenv('DYLD_LIBRARY_PATH', [binDir ':' currentLibPath]);
     end
+end
+
+function setDoNotUpdateFlag
+    full_path = mfilename('fullpath');
+    [path_root, ~, ~] = fileparts(full_path);    
+    filename = fullfile(path_root, 'do-not-update');
+    fileHandle = fopen(filename, 'w');
+    fclose(fileHandle);
+end
+
+function doNotUpdate = checkDoNotUpdate
+    full_path = mfilename('fullpath');
+    [path_root, ~, ~] = fileparts(full_path);
+    filename = fullfile(path_root, 'do-not-update');
+    doNotUpdate = (2 == exist(filename, 'file'));
+end
+
+function clearDoNotUpdate
+    full_path = mfilename('fullpath');
+    [path_root, ~, ~] = fileparts(full_path);
+    filename = fullfile(path_root, 'do-not-update');
+    if (2 == exist(filename, 'file'))
+        delete(filename);
+    end
+end
+
+function repoExists = checkIfGitRepoExists
+    full_path = mfilename('fullpath');
+    [path_root, ~, ~] = fileparts(full_path);
+    filename = fullfile(path_root, '.git');
+    repoExists = (7 == exist(filename, 'dir'));
 end
